requirements.txt
# Flask web framework - Latest stable release
Flask==3.1.1
Werkzeug==3.1.3

# ORM and database management
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-Migrate==4.0.7

# Database connectors

# Additional utilities
python-dotenv==1.0.1  # Environment variable management
gunicorn==23.0.0  # WSGI server for production

# Flask 3.1.1 compatible dependencies
Jinja2==3.1.4
MarkupSafe==3.0.2
Click==8.1.7
itsdangerous==2.2.0
